requirements/base.txt
django
django-environ
django-extensions
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2